+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [dev] - unreleased
 ### Added
+* BridgeDB supporting conversion of several database IDs
 * ComputeConverter class for conversions based on computation instead of querying 
 * ConverterBuilder which validates and initialises converters
 * reintroduced PubChem service using direct REST web interface